fix: 검색 단어 타입 분리
