+++ src/hooks/query/useWordSearch.ts
@@ -1,20 +1,6 @@
 import QUERY_KEYS from '@/constants/queryKey';
 import { useSuspenseQuery } from '@tanstack/react-query';
-
-export type SearchWordData = {
-  id: string;
-  name: string;
-  description: string;
-  diacritic: string[];
-  pronunciation: string[];
-  wordLike: boolean;
-};
-
-export type SearchWord = {
-  data: {
-    data: SearchWordData[];
-  };
-};
+import { SearchWord } from '@/types/search';
 
 export const getSearchWord = async (wordName: string) => {
   const res = await fetch(
